Add buttons to copy ffmpeg command line
Both lists of broadcasts (Following and Following2) can have buttons to generate and copy to clipboard a command to start ffmpeg with all necessary arguments to download broadcast.
Enable buttons in setting if you need them.
The the name of executable and additional arguments can also be changed in settings if needed.